requirements.txt
amqp==5.3.1
anyio==4.11.0
asgiref==3.11.0
billiard==4.2.3
build==1.3.0
CacheControl==0.14.4
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crashtest==0.4.1
cron_descriptor==2.0.6
distlib==0.4.0
Django==5.2.8
django-celery-beat==2.8.1
django-filter==25.2
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
dulwich==0.24.10
fastjsonschema==2.21.2
filelock==3.20.0
findpython==0.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.7.0
kombu==5.5.4
more-itertools==10.8.0
msgpack==1.1.2
packaging==25.0
pbs-installer==2025.11.20
pillow==12.0.0
pkginfo==1.12.1.2
platformdirs==4.5.0
poetry==2.2.1
poetry-core==2.2.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.3
trove-classifiers==2025.11.14.15
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.35.4
wcwidth==0.2.14
xattr==1.3.0
zstandard==0.25.0
